scripts/python_shit/pyproject.toml
[project]
name = "redis-supabase-app"
version = "0.1.0"
description = "FastAPI app with Redis and Supabase"
dependencies = [
    "fastapi>=0.100.0",
    "uvicorn>=0.22.0",
    "redis>=4.5.0",
    "supabase>=1.0.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
]